+++ src/main/java/com/dayaeyak/user/domain/user/UserAdminController.java
@@ -1,6 +1,6 @@
 package com.dayaeyak.user.domain.user;
 
-import com.dayaeyak.user.common.constraints.UserResponseMessage;
+import com.dayaeyak.user.common.constant.UserResponseMessage;
 import com.dayaeyak.user.common.entity.ApiResponse;
 import com.dayaeyak.user.domain.user.dto.request.UserAdminUpdateRequestDto;
 import com.dayaeyak.user.domain.user.dto.response.UserSearchPageResponseDto;
